:rocket:
Removed debug and combined with log.
Renamed mrdoc.js to mrdoc.
Removed options from local utilities and moved it to mr-doc-utils.
