+++ bin/mrdoc
@@ -1,18 +1,20 @@
 #!/usr/bin/env node
 /* eslint-env node */
 'use strict';
-
 const Option = require('mr-doc-utils').Option;
-const debug = require('../src/utils/debug');
+const Log = require('../src/utils/log');
 const CLI = require('../src/cli');
 
-// Parse the arguments
+// Parse the arguments.
 const argv = CLI.parse();
+// Get log level.
 const level = argv.level || argv.l;
 // Set up the logger.
-debug(Option.log({
+const log = new Log(Option.log({
   level: level !== 'silent' ? level : ['info', 'warn'],
   silent: level === 'silent',
 }));
+// DEBUG: Launch.
+log.debug(log.color.blue('Launching Mr. Doc CLI.'));
 // Launch the CLI.
 CLI.launch(argv);
